chore(panelRegistry): extract addPanel and hydration batch from core.ts
- Move hydration batch singleton + helpers to hydrationBatch.ts
- Move addPanel + private grid/non-trash counters to addPanel.ts
- core.ts shrinks from 1080 to 408 lines; logic preserved verbatim
- index.ts wires createAddPanelActions and re-exports isHydrationBatchActive
  from the new home (consumers in panelStore unchanged)
